change structure of project and rewrite cmake
