+++ CMakeLists.txt
@@ -3,4 +3,4 @@ cmake_minimum_required(VERSION 3.10.2)
 project(Eyes_Protection_Reminder VERSION 4.1.0)
 
 add_subdirectory(libs/wxWidgets)
-add_subdirectory(src)
+add_subdirectory(source)
